feat: Database for tags implemented #29
This commit adds the models and migrations for forum tags
